requirements.txt
pymongo
python-dotenv==0.6.4
python-telegram-bot==6.1.0
psycopg2==2.7.1
